Refactor: Moved UserInput from UI to Helper.
